+++ CHANGELOG.md
@@ -1576,3 +1576,4 @@
 - [2026-05-29 12:04] feat(analytics): add bi-directional event bus handler for analytics
 - [2026-05-29 12:32] feat(analytics): add async analytics stream provider
 - [2026-05-29 12:34] test(test): implement widget integration tests for test
+- [2026-05-29 12:46] feat(i18n): integrate custom telemetry logger in i18n